[Code/CppStandard] Massive cleanup warnings (#388)
* Warnings cleanup and modernize code

* Massive warning removal

* Massive warning purge

* Warning purge AND fixed script flag assignement wrong

* Revert all context used + warnings purge => compiled ok

* Boost deprecated warning

* PathGenerator warning clean

* More warning removal

* Boost 1.87 deprecated fix

* Boost 1.87 include removed for later

* Unused include removal + more warnings cleaning

* More warnings

* Includes

* Last includes

* Fix for GCC workflow:
/home/runner/work/Legends-of-Azeroth-Pandaria-5.4.8/Legends-of-Azeroth-Pandaria-5.4.8/src/server/game/Handlers/NPCHandler.cpp:500:65: error: invalid use of incomplete type ‘class CreatureAI’
  500 |     if (!sScriptMgr->OnGossipHello(_player, unit) && !unit->AI()->OnGossipHello(_player))
      |                                                                 ^~

---------

+++ src/server/game/Battlefield/Battlefield.cpp
@@ -21,8 +21,6 @@
 #include "CreatureTextMgr.h"
 #include "GridNotifiers.h"
 #include "GridNotifiersImpl.h"
-#include "GridNotifiers.h"
-#include "GridNotifiersImpl.h"
 #include "Group.h"
 #include "GroupMgr.h"
 #include "Map.h"
@@ -1028,7 +1026,7 @@ bool BfCapturePoint::Update(uint32 diff)
     }
 
     // get the difference of numbers
-    float fact_diff = ((float) m_activePlayers[0].size() - (float) m_activePlayers[1].size()) * diff / BATTLEFIELD_OBJECTIVE_UPDATE_INTERVAL;
+    float fact_diff = ((float) m_activePlayers[0].size() - (float) m_activePlayers[1].size()) * (float)diff / (float)BATTLEFIELD_OBJECTIVE_UPDATE_INTERVAL;
     if (G3D::fuzzyEq(fact_diff, 0.0f))
         return false;
